Refactor FolderCompletion and FolderSummary types for ...EventData.
Move into package lib/model/types.
